+++ doc/CHANGELOG.md
@@ -12,6 +12,10 @@
 - Added docker service
 - Added a sweet of scripts to help with managing the set up and images
 - Added github action pipeline to check changelog in PR's
+- Added World file downloads
+- Added backend testing framework
+- Added different container port to app port configs for games
+- Added valheim manifest
 
 ### Fixed
 - Dockerfiles updated and docker compose aswell.